[ty] Move method-related types to a submodule
